Flytter håndtering av dokumentasjon og vedlegg ut fra AppContext.ts og tilbake til Dokumentasjon.tsx
